rename realHostTriple -> rustTriple
This function's entire job is to convert the nix triple to a rust triple.
"real" was confusing.  "Real" to what?  It's also not always used to map a host
triple and sometimes was used to map a build triple.

The world is a bit better now.

Second, the generation of the synthetic Cargo.toml used to build each crate drv
in mkcrate.nix was using a combination of rustTriple and the explicit target
override.  That has been slightly normalized.

There is still a problem in pkgsBuild in make-package-set/simple.nix.  I don't
think the explicit target we set for the host output needs to even touch the
depsBuildBuild outputs.

The detection for when to enable multiple targets when generating the Cargo.toml
for build was changed to compare the two triples we generate.  This means even
if crossSystem is not set but the explicit target is used, the difference will
be picked up and both target entries will be present in the toml.

I'm concerned about the CC flags becuase it may mean we provide a triple that
Rust understands and the cc cannot.  This is especially true if the explicit
target is provided by the user.