! erase this from history
I've been putting the explicit target into the nixpkgs call.  That's a bit of a
dangerous pitfall.  Might rename the argument just to cause nixpkgs to fail.

This is really close to working this way other than the fact that I can't set
the crossSystem without oxalica trying to get too many components for the Rust toolchain.